Refonte des composant avec les item en étape et la recherche dans un nouveau composant
